+++ src/main/java/Coin.java
@@ -21,7 +21,7 @@ public class Coin extends PowerUp {
 
     @Override
     public Rectangle getBounds() {
-        return new Rectangle(x, y, w, h);
+        return new Rectangle((int)x, (int)y, w, h);
     }
 
 }
