+++ src/main/java/GameObject.java
@@ -4,12 +4,12 @@ public interface GameObject {
     void update();
     void paint(Graphics g);
     Rectangle getBounds();
-    int getX();
-    void setX(int x);
-    int getY();
-    void setY(int y);
-    int getSpeedX();
-    void setSpeedX(int sx);
-    int getSpeedY();
-    void setSpeedY(int sy);
+    double getX();
+    void setX(double x);
+    double getY();
+    void setY(double y);
+    double getSpeedX();
+    void setSpeedX(double sx);
+    double getSpeedY();
+    void setSpeedY(double sy);
 }
